core_backend/requirements.txt
Flask==2.3.3
PyMongo==4.5.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
PyJWT==2.8.0
werkzeug==3.1.3
pytz==2025.2
flask-cors==6.0.1
